improve functions.phpを作成
