+++ delete.php
@@ -1,21 +1,10 @@
 <?php
 session_start();
 require('dbconnect.php');
+require_once('functions.php');
 
-//0のみは弾く
-function v1($value) {
-    return preg_match('/\A[0]+\z/',$value);
-}
-//頭に0は弾く
-function v2($value) {
-    return preg_match('/\A[0]+[0-9]+\z/',$value);
-}
-//3桁以上は弾く
-function v3($value) {
-    return preg_match('/\A[1-9][0-9]{3,}\z/',$value);
-}
 
-//(res(返信))のパラメータ値チェック
+//idのパラメータ値チェック
 if(isset($_REQUEST['id'])){
     $id_ck = mb_convert_kana($_REQUEST['id'], 'n', 'UTF-8');
     if(v1($id_ck) || v2($id_ck) || v3($id_ck) || !ctype_digit($id_ck)){
