+++ rtiine.php
@@ -1,20 +1,8 @@
 <?php
 session_start();
 require('dbconnect.php');
+require_once('functions.php');
 
-
-//0のみは弾く
-function v1($value) {
-    return preg_match('/\A[0]+\z/',$value);
-}
-//頭に0は弾く
-function v2($value) {
-    return preg_match('/\A[0]+[0-9]+\z/',$value);
-}
-//3桁以上は弾く
-function v3($value) {
-    return preg_match('/\A[1-9][0-9]{3,}\z/',$value);
-}
 
 //「リツイートいいね」パラメータ値のチェック
 if(isset($_REQUEST['ine'])){
